refactor: User, UserBase and controllers
