+++ Http/Controllers/Api/UsersController.php
@@ -8,8 +8,8 @@ use Illuminate\Http\Response;
 use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\Hash;
 use WebReinvent\VaahCms\Entities\Registration;
-use WebReinvent\VaahCms\Entities\User;
 use WebReinvent\VaahCms\Models\Permission;
+use WebReinvent\VaahCms\Models\User;
 
 class UsersController extends Controller
 {
