+++ Http/Controllers/Settings/SettingsController.php
@@ -5,7 +5,7 @@ namespace WebReinvent\VaahCms\Http\Controllers\Settings;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Routing\Controller;
-use WebReinvent\VaahCms\Entities\User;
+use WebReinvent\VaahCms\Models\User;
 
 class SettingsController extends Controller
 {
