+++ Models/Role.php
@@ -6,8 +6,8 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Str;
 use WebReinvent\VaahCms\Traits\CrudWithUuidObservantTrait;
-use WebReinvent\VaahCms\Entities\User;
 use WebReinvent\VaahCms\Models\Permission;
+use WebReinvent\VaahCms\Models\User;
 
 class Role extends RoleBase
 {
