+++ Models/User.php
@@ -6,7 +6,7 @@ use DateTimeInterface;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Str;
-use WebReinvent\VaahCms\Entities\Role;
+use WebReinvent\VaahCms\Models\Role;
 use WebReinvent\VaahCms\Traits\CrudWithUuidObservantTrait;
 
 class User extends UserBase
@@ -53,6 +53,7 @@ class User extends UserBase
     ];
     //-------------------------------------------------
     protected $appends = [
+        'name'
     ];
 
     //-------------------------------------------------
@@ -94,6 +95,28 @@ class User extends UserBase
             ->getColumnListing($this->getTable());
     }
 
+    //-------------------------------------------------
+    public function getNameAttribute() {
+
+        if($this->display_name)
+        {
+            return $this->display_name;
+        }
+
+        $name = $this->first_name;
+
+        if($this->middle_name)
+        {
+            $name .= " ".$this->middle_name;
+        }
+
+        if($this->last_name)
+        {
+            $name .= " ".$this->last_name;
+        }
+
+        return $name;
+    }
     //-------------------------------------------------
     public function scopeExclude($query, $columns)
     {
