+++ Models/UserBase.php
@@ -304,7 +304,7 @@ class UserBase extends Authenticatable
     //-------------------------------------------------
     public function roles()
     {
-        return $this->belongsToMany('WebReinvent\VaahCms\Entities\Role',
+        return $this->belongsToMany(Role::class,
             'vh_user_roles', 'vh_user_id', 'vh_role_id'
         )->withPivot('is_active',
             'created_by',
