+++ Routes/backend/routes-users.php
@@ -6,7 +6,7 @@ Route::group(
 
         //'middleware' => ['web', 'has.backend.access'],
 
-        'namespace' => 'WebReinvent\VaahCms\Http\Controllers',
+        'namespace' => 'WebReinvent\VaahCms\Http\Controllers\Backend',
     ],
     function () {
         /**
